Add dot-zshrc
This file is meant to be symlinked to ~/.zshrc

An optional ~/.zshrc-config file is supported to provide for local config,
should any be necessary. This file is read first, before anything else
happens.

The ~/.zshrc symlink is read to find the location of the zshrc directory,
and then files from $zshrcdir/rc/* are read in to setup the environment.
Currently, no rc/* files exist (will be done in future commits).